Move bitflags struct to lexer
